requirements.txt
seaborn
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
numpy==1.16.5
pandas==0.25.1
protobuf==3.8.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
Werkzeug==0.15.5
wrapt==1.11.2
xlrd==1.2.0
